# Speed up func calls