--- a/bench/func_call.rb
+++ b/bench/func_call.rb
@@ -4,15 +4,16 @@ Bench.ips do |x|
   engine = Wasmtime::Engine.new
   linker = Wasmtime::Linker.new(engine)
   mod = Wasmtime::Module.from_file(engine, "examples/gcd.wat")
+  store = Wasmtime::Store.new(engine)
+  instance = linker.instantiate(store, mod)
+  func = instance.export("gcd").to_func
 
   x.report("Instance#invoke") do
-    store = Wasmtime::Store.new(engine)
-    linker.instantiate(store, mod).invoke("gcd", 5, 1)
+    instance.invoke("gcd", 5, 1)
   end
 
   x.report("Func#call") do
-    store = Wasmtime::Store.new(engine)
-    linker.instantiate(store, mod).export("gcd").to_func.call(5, 1)
+    func.call(5, 1)
   end
 
   x.compare!